# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AED Foreigner API",
        "description": "API documentation for AED Foreigner Enrollment",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/signing-identities": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "List Signing Identities",
                "operationId": "getSigningIdentities",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/signing-identities/{identityId}": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Get Signing Identity Details",
                "operationId": "getSigningIdentity",
                "parameters": [
                    {
                        "name": "identityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Delete Signing Identity",
                "operationId": "deleteSigningIdentity",
                "parameters": [
                    {
                        "name": "identityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/signing-identities/{identityId}/status": {
            "put": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Update Signing Identity Status",
                "operationId": "updateSigningIdentityStatus",
                "parameters": [
                    {
                        "name": "identityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value",
                                    "reason"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "string"
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/signing-identities/provision": {
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Provision Certificate (Signature)",
                "description": "Starts the certificate issuance process on TrustedX RAP.",
                "operationId": "provisionSignature",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "type",
                                    "subscription_id"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "citizen",
                                            "employee"
                                        ]
                                    },
                                    "subscription_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Process finished or in progress"
                    },
                    "400": {
                        "description": "Validation or Access error"
                    },
                    "403": {
                        "description": "Forbidden (Manager approval or account inactive)"
                    }
                }
            }
        },
...
    },
...
}